refactor: rename no-unknown-export rule
