+++ packages/eslint/src/index.js
@@ -3,7 +3,7 @@ export rules from "./rules";
 export const configs = {
   recommended: {
     rules: {
-      "webassembly/no-unexisting-export": 2 // error
+      "webassembly/no-unknown-export": 2 // error
     }
   }
 };
